fix(achieve&training) : fix achieve&training
1. 회피 훈련 쿨타임감소로 변경
2. 업적 단위 추가

+++ Assets/Scripts/Unit/LobbyPlayer/Info/LobbyPlayerInfo.cs
@@ -177,8 +177,8 @@ public class LobbyPlayerInfo
         }
     }
     [SerializeField]
-    private int mTrainingDodgeTime;
-    public int TrainingDodgeTime
+    private float mTrainingDodgeTime;
+    public float TrainingDodgeTime
     {
         get { return mTrainingDodgeTime; }
         set
